Refactor CI and documentation notebook imports for better modularity
